Rename ApiResource to ApiHttpClient
