+++ workspace/apps/pidp/src/app/features/admin/shared/resources/admin-resource.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 
 import { Observable, catchError, of } from 'rxjs';
 
-import { ApiResource } from '@app/core/resources/api-resource.service';
+import { ApiHttpClient } from '@app/core/resources/api-http-client.service';
 
 export interface PartyList {
   id: number;
@@ -17,7 +17,7 @@ export interface PartyList {
   providedIn: 'root',
 })
 export class AdminResource {
-  public constructor(private apiResource: ApiResource) {}
+  public constructor(private apiResource: ApiHttpClient) {}
 
   public getParties(): Observable<PartyList[]> {
     return this.apiResource.get<PartyList[]>('/admin/parties').pipe(
