+++ workspace/libs/shared/data-access/src/lib/interfaces/crud-resource.interface.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs';
 
-import { NoContent } from '../classes/abstract-http-resource';
+import { NoContent } from '../classes/abstract-http-client';
 
 export interface ICrudResource<T> {
   create(id: number, payload: T): Observable<T | null>;
