+++ workspace/libs/shared/data-access/src/lib/interfaces/index.ts
@@ -1,2 +1,2 @@
 export * from './crud-resource.interface';
-export * from './http-resource.interface';
+export * from './http-client.interface';
